refactorización del 90% de las rutas y endpoints

+++ app/adapters/output/orm/models/job_offer_required_skill_model.py
@@ -1,7 +1,8 @@
 from sqlalchemy import Column, Integer, ForeignKey
 from app.infraestructure.database.base import Base
+from sqlalchemy.dialects.mysql import SMALLINT
 
 class JobOfferRequiredSkillModel(Base):
     __tablename__ = 'job_offer_required_skill'
     job_offer_id = Column(Integer, ForeignKey('job_offer.id'), primary_key=True, nullable=False)
-    skill_id = Column(Integer, ForeignKey('skill.id'), primary_key=True, nullable=False)
+    skill_id = Column(SMALLINT, ForeignKey('skill.id'), primary_key=True, nullable=False)
